requirements.txt
# API Dependencies
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
tensorflow==2.17.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
scikit-learn==1.5.2
matplotlib==3.9.2
pydantic==2.9.0
requests==2.32.3
python-dotenv==1.0.1

